+++ src/main/java/com/smartsolutions/eschool/student/facade/StudentFeePaymentsFacade.java
@@ -8,6 +8,8 @@ import jakarta.validation.Valid;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Component
 @Scope("prototype")
 public class StudentFeePaymentsFacade {
@@ -22,11 +24,11 @@ public class StudentFeePaymentsFacade {
     public StudentFeePaymentRequestDTO studentFeePayment(Long studentId, @Valid StudentFeePaymentRequestDTO requestDTO) {
         return studentFeePaymentsService.studentFeePayment(studentId, requestDTO);
     }
-    public Double getTotalFeeCollected(Long academicYearId) {
+    public BigDecimal getTotalFeeCollected(Long academicYearId) {
         return studentFeePaymentsService.getTotalFeeCollected(academicYearId);
     }
 
-    public Double getCollectedUpToCurrentMonth(){
+    public BigDecimal getCollectedUpToCurrentMonth(){
         return studentFeePaymentsService.getCollectedUpToCurrentMonth();
     }
 }
